Radial functions

These functions are purely radial, in which case only a quadrant is computed and then copied. They are all based on a separable view version of rr2 but then continue to calculate on a corner of a grid, which is (as a last step) replicated and mirrored.

SeparableFunctions.calc_radial_symm — Function

calc_radial_symm!([::Type{TA},] sz::NTuple,  fct; scale = one(real(T)), myrr2sep = rr2_sep(size(arr); scale=scale, offset=size(arr).÷2 .+1)) where {N,T}

evaluates the radial function fct over the entire array. The function needs to accept the radius as argument. The calculation is done fast by only evaluating on the first quadrant and replicating the results by copy operations using copy_corners!().

#Arguments

TA: The array type for the newly created array.

sz: The size of the newly created array.

fct: The function of the radius, to be evaluate on the array coordinates.

scale: the vetorized scaling of the pixels (only used, if myrr2sep is not supplied by the user)

myrr2sep: The separable xx^2 and yy^2 etc. information as obtained by rr2_sep().

SeparableFunctions.calc_radial_symm! — Function

calc_radial_symm!(arr::TA, fct; scale = one(real(T)), myrr2sep = rr2_sep(size(arr); scale=scale, offset=size(arr).÷2 .+1)) where {N,T}

evaluates the radial function fct over the entire array. The function needs to accept the radius as argument. The calculation is done fast by only evaluating on the first quadrant and replicating the results by copy operations using copy_corners!().

#Arguments

arr: The array into which to evaluate the radial function

fct: The function of the radius, to be evaluate on the array coordinates.

scale: the vetorized scaling of the pixels (only used, if myrr2sep is not supplied by the user)

myrr2sep: The separable xx^2 and yy^2 etc. information as obtained by rr2_sep().

SeparableFunctions.propagator_col — Function

propagator_col([]::Type{TA},] sz::NTuple{N, Int}; Δz=one(eltype(TA)), k_max=0.5f0, scale=0.5f0 ./ (max.(sz ./ 2, 1))) where{TA, N}

generates a propagator for propagating optical fields via exp(i kz Δz) with kz=sqrt(k0^2-kx^2-ky^2). The k-space radius is stated by k_max relative to the Nyquist frequency, as long as the scale remains to be 1 ./ (2 max.(sz ./ 2, 1))).

#Arguments

TA: type of the array to generate. E.g. Array{Float64} or CuArray{Float32}.

sz: size of the array to generate. If a 3rd dimension is present, a stack a propagators is returned, one for each multiple of Δz.

Δz: distance in Z to propagate per slice.

k_max: maximum propagation radius in k-space. I.e. limit of the k-sphere. This is not the aperture limit!

scale: specifies how to interpret k-space positions. Should remain to be 1 ./ (2 max.(sz ./ 2, 1))).

SeparableFunctions.propagator_col! — Function

propagator_col(arr::AbstractArray{T,N}; Δz=one(eltype(TA)), k_max=0.5f0, scale=0.5f0 ./ (max.(sz ./ 2, 1))) where{TA, N}

generates a propagator for propagating optical fields via exp(i kz Δz) with kz=sqrt(k0^2-kx^2-ky^2). The k-space radius is stated by k_max relative to the Nyquist frequency, as long as the scale remains to be 1 ./ (2 max.(sz ./ 2, 1))).

#Arguments

arr: the array to fill with propagators. If a 3rd dimension is present, a stack a propagators is returned, one for each multiple of Δz.

Δz: distance in Z to propagate per slice.

k_max: maximum propagation radius in k-space. I.e. limit of the k-sphere. This is not the aperture limit!

scale: specifies how to interpret k-space positions. Should remain to be 1 ./ (2 max.(sz ./ 2, 1))).

SeparableFunctions.phase_kz_col — Function

phase_kz_col([::Type{TA},] sz::NTuple{N, Int};k_max=0.5f0, scale=0.5f0 ./ (max.(sz ./ 2, 1))) where{TA, N}

Calculates a propagation phase (without the 2pi factor!) for a given z-position, which can be defined via a 3rd entry in the offset supplied to the function. By default, Nyquist sampling it is assumed such that the lateral kxy corresponds to the XY border in frequency space at the edge of the Ewald circle. However, via the xy scale entries the kmax can be set appropriately. The propagation equation should Δz .* sqrt.(1-kxyrel^2) as the propagation phase. The Z-propagation distance (Δz) has to be specified in units of the wavelength in the medium (λ = n*λ₀). Note that since the phase is normalized to 1 instead of 2pi, you need to use this phase in the following sense: `cispi.(2.*phasekz(...))`.

#Arguments

TA: Array type of the result array. For cuda calculations use CuArray{Float32}.

sz: Size (2D) of the result array.

k_max: maximum propagation radius in k-space. I.e. limit of the k-sphere. This is not the aperture limit!

scale: specifies how to interpret k-space positions. Should remain to be 1 ./ (2 max.(sz ./ 2, 1))).

SeparableFunctions.phase_kz_col! — Function

phase_kz_col!(arr::AbstractArray{T,N}; k_max=0.5f0, scale=0.5f0 ./ (max.(sz ./ 2, 1))) where{TA, N}

Calculates a propagation phase (without the 2pi factor!) for a given z-position, which can be defined via a 3rd entry in the offset supplied to the function. By default, Nyquist sampling it is assumed such that the lateral kxy corresponds to the XY border in frequency space at the edge of the Ewald circle. However, via the xy scale entries the kmax can be set appropriately. The propagation equation uses Δz .* sqrt.(1-kxyrel^2) as the propagation phase. The Z-propagation distance (Δz) has to be specified in units of the wavelength in the medium (λ = n*λ₀). Note that since the phase is normalized to 1 instead of 2pi, you need to use this phase in the following sense: `cispi.(2.*phasekz(...))`.

#Arguments

arr: the array to fill with propagators. If a 3rd dimension is present, a stack a propagators is returned, one for each multiple of Δz.

k_max: maximum propagation radius in k-space. I.e. limit of the k-sphere. This is not the aperture limit!

scale: specifies how to interpret k-space positions. Should remain to be 1 ./ (2 max.(sz ./ 2, 1))).