Utility functions

Helper functions

SeparableFunctions.arg_n — Function

arg_n(n, args...)

returns a modified version of args, such that scalars remain scalar but in vectors the n'th position is picked. This is useful for calling separable functions with their scalar arguments differing for each vector entry.

Example

jdoctest julia> args = (1,(4,5)) (1, (4, 5)) julia> collect(SeparableFunctions.arg_n(2, args)) 2-element Vector{Int64}: 1 5

SeparableFunctions.kwarg_n — Function

kwarg_n(n, args...)

returns a modified version of keyword-args kwargs, such that scalar values remain scalar but in vectors the n'th position is picked. This is useful for calling separable functions with their scalar arguments differing for each vector entry.

# Example

julia> kw = (a=1,b=(4,5))
(a = 1, b = (4, 5))
julia> SeparableFunctions.kwarg_n(2, kw)
(a = 1, b = 5)

SeparableFunctions.pick_n — Function

pick_n(n, v)

picks the nth value of the vector v or return the scalar v.

SeparableFunctions.get_corner_ranges — Function

get_corner_ranges(sz::NTuple{N}; shifted_dims = zeros(Bool, N), inv_dims = zeros(Bool, N), full_dims=zeros(Bool, N)) where {N}

returns a tuple of ranges that can be used for indexing various corners of an N-dimensional dataset. #Arguments

sz: total size of the data

shifted_dims: an iterable of Boolean defining which dimension to shift (by half the datasize)

inv_dims: an iterable of Boolean defining which which range dimension

full_dims: an iterable of Boolean defining which dimension to include fully as a colon (:)

SeparableFunctions.copy_last_dim! — Function

copy_last_dim(arr::AbstractArray{T,N}) where{T,N}

mirrors the last dimension exploiting the 1D-nature of the trailing dimension using 1D indexing of the ND array. However even-size dimensions need special treatment.

Missing docstring.

Missing docstring for SeparableFunctions.similar_arr_type. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeparableFunctions.get_real_arr_type. Check Documenter's build log for details.