SeparableFunctions.jl

SeparableFunctions Interface

The abstract SeparableFunctions definition

SeparableFunctions — Module

Calculates multidimensional functions faster by exploiting their separability. Often a function involves an operation such as a (complex) exponential which by itself is computationally relatively heavy. Yet a number of multidimenstional functions are separable, which means that they can be written as a product (or sum) of single-dimensional functions. A good example of a separable function is a Gaussian function, which can be written as a product of a purely X-dependend Gaussian with a purely Y-dependend Gaussian.

In this package, multidimensional functions are computed by first calculating their single-dimensional values and then creating the final multidimensional result by an outer product. Since multiplications and the broadcasting mechanism of Julia are fast compared to the evaluation of the function at each multidimensional position, the final result is calculated faster. The typical speedup can be an order of magnitude.

The package offers a general way of calculating separable functions as well as a LazyArrays version of that function which can then be used inside other expressions. The non-lazy version should currently also work with CUDA.jl, however the LazyArrays version does not. To nevertheless use separable expressions in CUDA.jl, you can reside to externally applying the broadcast operator to the separable expression (see the gaussian_sep example below).

The package further offers a number of predifined separable function implementations such as gaussian_col() collects a multidimensional array of a multidimensional Gaussian via a fast seperable implementation, gaussian_lz() yields a lazy representation via LazyArrays and sep = gaussian_sep() yields an iterable of separable pre-oriented vectors which can easily be mutually combined via .*(sep...).

Another noteworthy example is the complex plain wave as represented by the respective function exp_ikx_col(), exp_ikx_lz(), exp_ikx_sep().

All separable functions share a common interface with Standard non-named arguments

the first optional argument being the type of the result array. Examples are Array{Float64}, CuArray{Float32} and the default depends on the function but uses a 32-bit result type where applicable.

the next argument is the size of the result data supplied as a tuple.

optionally a further argument can specify the position of zero of the array with respect to offset as given below. This allows for convenient N-dimensional shifting of the functions.

Named arguments:

offset: is a optional named argument specifying center of the array. By default the Fourier center, size().÷2 .+ 1 is chosen.

scale: multiplies the axes with these factors. This can be interpreted as a pixelsize.

Some functions have additional named arguments. E.g. gaussian has the additional named argument sigma specifying the width of the gaussian, even though this can also be controlled via scale. Note that this nomenclature is in large parts identical to the package IndexFunArrays.jl.

In general arguments can be supplied as single scalar values or vectors. If a scalar value is supplied, it will automatically be replicated as a vector. E.g. sigma=10.0 for a 2-dimensional array will be interpreted as sigma=(10.0, 10.0).

Here you can find the docstrings of currently implemented functions.

SeparableFunctions.calc_radial_symm! — Function

calc_radial_symm!(arr::TA, fct; scale = one(real(T)), myrr2sep = rr2_sep(size(arr); scale=scale, offset=size(arr).÷2 .+1)) where {N,T}

evaluates the radial function fct over the entire array. The function needs to accept the square of the radius as argument! The calculation is done fast by only evaluating on the first quadrant and replicating the results by copy operations using copy_corners!().

#Arguments

arr: The array into which to evaluate the radial function

fct: The radial function of the squared radius, to be evaluate on the array coordinates.

scale: the vetorized scaling of the pixels (only used, if myrr2sep is not supplied by the user)

myrr2sep: The separable xx^2 and yy^2 etc. information as obtained by rr2_sep().

SeparableFunctions.copy_corners! — Function

copy_corners!(arr::AbstractArray{T,N}) where {T,N}

replicates the first N-dimensional quadrant my several mirror operations over the entire array. The overwrites the entire array with the (mirrored) content of the first quadrant!

#Arguments

arr: The array in which the copy operations are performed

speedup_last_dim=true: if true a one-dimensional assignment trick is used for the last non-singleton dimension.

SeparableFunctions.separable_view — Function

separable_view{N}(fct, sz, args...; pos=zero(real(eltype(AT))), offset =  sz.÷2 .+1, scale = one(real(eltype(AT))), operation = .*)

creates an array view of an N-dimensional separable function. Note that this view consumes much less memory than a full allocation of the collected result. Note also that an N-dimensional calculation expression may be much slower than this view reprentation of a product of N one-dimensional arrays. See the example below.

Arguments:

fct: The separable function, with a number of arguments corresponding to length(args), each corresponding to a vector of separable inputs of length N. The first argument of this function is a Tuple corresponding the centered indices.

sz: The size of the N-dimensional array to create

args...: a list of arguments, each being an N-dimensional vector

offset: position of the center from which the position is measured

scale: defines the pixel size as vector or scalar. Default: 1.0.

operation: the separable operation connecting the separable dimensions

Example:

julia> fct = (r, sz, pos, sigma)-> exp(-(r-pos)^2/(2*sigma^2))
julia> my_gaussian = separable_view(fct, (6,5), (0.1,0.2), (0.5,1.0))
(6-element Vector{Float64}) .* (1×5 Matrix{Float64}):
 3.99823e-10  2.18861e-9   4.40732e-9   3.26502e-9   8.89822e-10
 1.3138e-5    7.19168e-5   0.000144823  0.000107287  2.92392e-5
 0.00790705   0.0432828    0.0871609    0.0645703    0.0175975
 0.0871609    0.477114     0.960789     0.71177      0.19398
 0.0175975    0.0963276    0.19398      0.143704     0.0391639
 6.50731e-5   0.000356206  0.000717312  0.000531398  0.000144823

SeparableFunctions.separable_create — Function

separable_create([::Type{TA},] fct, sz::NTuple{N, Int}, args...; pos=zero(real(eltype(AT))), offset =  sz.÷2 .+1, scale = one(real(eltype(AT))), operation = *, kwargs...) where {TA, N}

creates an array view of an N-dimensional separable function including memory allocation and collection. See the example below.

Arguments:

fct: The separable function, with a number of arguments corresponding to length(args), each corresponding to a vector of separable inputs of length N. The first argument of this function is a Tuple corresponding the centered indices.

sz: The size of the N-dimensional array to create

args...: a list of arguments, each being an N-dimensional vector

offset: position of the center from which the position is measured

scale: defines the pixel size as vector or scalar. Default: 1.0.

operation: the separable operation connecting the separable dimensions

Example:

julia> fct = (r, sz, sigma)-> exp(-r^2/(2*sigma^2))
julia> my_gaussian = separable_create(fct, (6,5), (0.5,1.0); pos=(0.1,0.2))
6×5 Matrix{Float32}:
 3.99823f-10  2.18861f-9   4.40732f-9   3.26502f-9   8.89822f-10
 1.3138f-5    7.19168f-5   0.000144823  0.000107287  2.92392f-5
 0.00790705   0.0432828    0.0871608    0.0645703    0.0175975
 0.0871608    0.477114     0.960789     0.71177      0.19398
 0.0175975    0.0963276    0.19398      0.143704     0.0391639
 6.50731f-5   0.000356206  0.000717312  0.000531398  0.000144823

<!–

SeparableFunctions Interface

SeparableFunctions.gaussian_col — Function

gaussian_col([::Type{TA},] sz::NTuple{N, Int}; sigma=ones(eltype(TA),N), pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

creates a multidimensional Gaussian by exployting separability speeding up the calculation.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

sigma: tuple of standard-deviation along each dimensional

returns the collected N-dimensional array.

#Example

julia> pos = (1.1, 0.2); sigma = (0.5, 1.0);
julia> my_gaussian = gaussian_col((6,5); sigma=sigma, pos=pos)
6×5 Matrix{Float32}:
    2.22857f-16  1.21991f-15  2.4566f-15   1.81989f-15  4.95978f-16
    3.99823f-10  2.18861f-9   4.40732f-9   3.26502f-9   8.89822f-10
    1.3138f-5    7.19168f-5   0.000144823  0.000107287  2.92392f-5
    0.00790705   0.0432828    0.0871608    0.0645703    0.0175975
    0.0871608    0.477114     0.960789     0.71177      0.19398
    0.0175975    0.0963276    0.19398      0.143704     0.0391639

SeparableFunctions.gaussian_lz — Function

gaussian_lz([::Type{TA},] sz::NTuple{N, Int}; sigma=ones(eltype(TA),N), pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

creates a multidimensional Gaussian by exployting separability speeding up the calculation.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

sigma: tuple of standard-deviation along each dimensional

returns the a lazy version of an N-dimensional array only using memory for the separated 1-dimensional arrays. See the _col version for an example.

SeparableFunctions.gaussian_sep — Function

gaussian_sep([::Type{TA},] sz::NTuple{N, Int}; sigma=ones(eltype(TA),N), pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

creates a multidimensional Gaussian by exployting separability speeding up the calculation.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

sigma: tuple of standard-deviation along each dimensional

returns an iterable with the one-dimensional arrays, which can be used via .*(res...) wiht res being the result of calling this function.

SeparableFunctions.normal_col — Function

normal_col([::Type{TA},] sz::NTuple{N, Int}; sigma=ones(eltype(TA),N), pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

creates a multidimensional normalized Gaussian by exployting separability speeding up the calculation. The integral of the multidimensional array, if it where infinite, is normalized to one.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

sigma: tuple of standard-deviation along each dimensional

returns the collected N-dimensional array. #Example

julia> pos = (1.1, 0.2); sigma = (0.5, 1.0);
julia> my_normal = normal_col((6,5); sigma=sigma, pos=pos)
6×5 Matrix{Float32}:
    7.09377f-17  3.88309f-16  7.81959f-16  5.79289f-16  1.57875f-16
    1.27268f-10  6.96656f-10  1.40289f-9   1.03929f-9   2.83239f-10
    4.18196f-6   2.28918f-5   4.60985f-5   3.41506f-5   9.30713f-6
    0.00251689   0.0137773    0.0277442    0.0205534    0.00560145
    0.0277442    0.15187      0.305829     0.226564     0.0617458
    0.00560145   0.030662     0.0617458    0.0457424    0.0124663

SeparableFunctions.normal_lz — Function

normal_lz([::Type{TA},] sz::NTuple{N, Int}; sigma=ones(eltype(TA),N), pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

creates a multidimensional normalized Gaussian by exployting separability speeding up the calculation. The integral of the multidimensional array, if it where infinite, is normalized to one.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

sigma: tuple of standard-deviation along each dimensional

returns the a lazy version of an N-dimensional array only using memory for the separated 1-dimensional arrays. See the _col version for an example.

SeparableFunctions.normal_sep — Function

normal_sep([::Type{TA},] sz::NTuple{N, Int}; sigma=ones(eltype(TA),N), pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

creates a multidimensional normalized Gaussian by exployting separability speeding up the calculation. The integral of the multidimensional array, if it where infinite, is normalized to one.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

sigma: tuple of standard-deviation along each dimensional

returns an iterable with the one-dimensional arrays, which can be used via .*(res...) wiht res being the result of calling this function.

SeparableFunctions.ramp_col — Function

ramp_col([::Type{TA},] sz::NTuple{N, Int}; slope, pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

creates an N-dimensional ramp along the gradient-direction defined by slope. Note that this disagrees with the nomenclature of the ramp-function defined in IndexFunArrays.jl.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

slope: a vector defining the N-dimensional gradient of the ramp.

returns the collected N-dimensional array. #Example julia> myramp = rampcol((6,5); slope=(0.0,0.5), pos=pos) 6×5 Matrix{Float32}: -1.1 -0.6 -0.1 0.4 0.9 -1.1 -0.6 -0.1 0.4 0.9 -1.1 -0.6 -0.1 0.4 0.9 -1.1 -0.6 -0.1 0.4 0.9 -1.1 -0.6 -0.1 0.4 0.9 -1.1 -0.6 -0.1 0.4 0.9 ```

SeparableFunctions.ramp_lz — Function

ramp_lz([::Type{TA},] sz::NTuple{N, Int}; slope, pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

creates an N-dimensional ramp along the gradient-direction defined by slope. Note that this disagrees with the nomenclature of the ramp-function defined in IndexFunArrays.jl.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

slope: a vector defining the N-dimensional gradient of the ramp.

returns the a lazy version of an N-dimensional array only using memory for the separated 1-dimensional arrays. See the _col version for an example.

SeparableFunctions.ramp_sep — Function

ramp_sep([::Type{TA},] sz::NTuple{N, Int}; slope, pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

creates an N-dimensional ramp along the gradient-direction defined by slope. Note that this disagrees with the nomenclature of the ramp-function defined in IndexFunArrays.jl.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

slope: a vector defining the N-dimensional gradient of the ramp.

returns an iterable with the one-dimensional arrays, which can be used via .*(res...) wiht res being the result of calling this function.

SeparableFunctions.rr2_col — Function

rr2_col([::Type{TA},] sz::NTuple{N, Int}; pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

yields the absolute square of the distance to the zero-position.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

returns the collected N-dimensional array. #Example

julia> pos = (1.1, 0.2); 
julia> my_rr2 = rr2_col((6,5); pos=pos)
6×5 Matrix{Float32}:
    21.65  18.25  16.85  17.45  20.05
    14.45  11.05   9.65  10.25  12.85
    9.25   5.85   4.45   5.05   7.65
    6.05   2.65   1.25   1.85   4.45
    4.85   1.45   0.05   0.65   3.25
    5.65   2.25   0.85   1.45   4.05

SeparableFunctions.rr2_lz — Function

rr2_lz([::Type{TA},] sz::NTuple{N, Int}; pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

yields the absolute square of the distance to the zero-position.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

returns the a lazy version of an N-dimensional array only using memory for the separated 1-dimensional arrays. See the _col version for an example.

SeparableFunctions.rr2_sep — Function

rr2_sep([::Type{TA},] sz::NTuple{N, Int}; pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

yields the absolute square of the distance to the zero-position.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

returns an iterable with the one-dimensional arrays, which can be used via .+(res...) wiht res being the result of calling this function.

SeparableFunctions.box_col — Function

box_col([::Type{TA},] sz::NTuple{N, Int}; boxsize=sz./2, pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

creates a Boolean box, being True inside and False outside. The side-length of the box can be defined via the argument boxsize.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

boxsize: a vector defining each sidelength of the box.

returns the collected N-dimensional array. #Example

julia> pos = (1.1, 0.2); 
julia> my_box = box_col((6,5); pos=pos)
6×5 BitMatrix:
 0  0  0  0  0
 0  0  0  0  0
 0  0  0  0  0
 0  1  1  1  0
 0  1  1  1  0
 0  1  1  1  0

SeparableFunctions.box_lz — Function

box_lz([::Type{TA},] sz::NTuple{N, Int}; boxsize=sz./2, pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

creates a Boolean box, being True inside and False outside. The side-length of the box can be defined via the argument boxsize.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

boxsize: a vector defining each sidelength of the box.

returns the a lazy version of an N-dimensional array only using memory for the separated 1-dimensional arrays. See the _col version for an example.

SeparableFunctions.box_sep — Function

box_sep([::Type{TA},] sz::NTuple{N, Int}; boxsize=sz./2, pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

creates a Boolean box, being True inside and False outside. The side-length of the box can be defined via the argument boxsize.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

boxsize: a vector defining each sidelength of the box.

returns an iterable with the one-dimensional arrays, which can be used via .*(res...) wiht res being the result of calling this function.

SeparableFunctions.sinc_col — Function

sinc_col([::Type{TA},] sz::NTuple{N, Int}; pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

yields the outer product of sinc functions. This corresponds to the diffraction pattern of a rectangular aperture. Note that it is not circularly symmetric.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

returns the collected N-dimensional array. #Example

julia> pos = (1.1, 0.2); 
julia> my_sinc = sinc_col((6,5); pos=pos)
6×5 Matrix{Float32}:
  0.0020403   -0.00374056   0.0224433   0.00561083  -0.0024937
 -0.00269847   0.00494719  -0.0296831  -0.00742078   0.00329812
  0.00398345  -0.00730299   0.0438179   0.0109545   -0.00486866
 -0.00760477   0.0139421   -0.0836524  -0.0209131    0.00929472
  0.0836524   -0.153363     0.920177    0.230044    -0.102242
  0.00929472  -0.0170403    0.102242    0.0255605   -0.0113602

SeparableFunctions.sinc_lz — Function

sinc_lz([::Type{TA},] sz::NTuple{N, Int}; pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

yields the outer product of sinc functions. This corresponds to the diffraction pattern of a rectangular aperture. Note that it is not circularly symmetric.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

returns the a lazy version of an N-dimensional array only using memory for the separated 1-dimensional arrays. See the _col version for an example.

SeparableFunctions.sinc_sep — Function

sinc_sep([::Type{TA},] sz::NTuple{N, Int}; pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

yields the outer product of sinc functions. This corresponds to the diffraction pattern of a rectangular aperture. Note that it is not circularly symmetric.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

returns an iterable with the one-dimensional arrays, which can be used via .*(res...) wiht res being the result of calling this function.

SeparableFunctions.exp_ikx_col — Function

exp_ikx_col([::Type{TA},] sz::NTuple{N, Int}; pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

yield an exp.(1im .* k .* Δx) function. A scaling of one means that the shift_by argument corresponds to Δx in integer pixels, but the term is in Fourier space. Note that the meaning of scale is different compared to the version in IndexFunArray.jl. Here the default scaling of 1.0 corresponds to the default scaling of ScaFT in IndexFunArrays.jl.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

shift_by: a vector defining the real-space shift that would be caused by this function being multiplied in Fourier-space.

returns the collected N-dimensional array. #Example

julia> pos = (1.1, 0.2); 
julia> my_exp_ikx = exp_ikx_col((6,5); shift_by=(1.0,0.0), pos=pos)
6×5 Matrix{ComplexF32}:
 -0.406737-0.913545im  -0.406737-0.913545im  -0.406737-0.913545im  -0.406737-0.913545im  -0.406737-0.913545im
 -0.994522-0.104528im  -0.994522-0.104528im  -0.994522-0.104528im  -0.994522-0.104528im  -0.994522-0.104528im
 -0.587785+0.809017im  -0.587785+0.809017im  -0.587785+0.809017im  -0.587785+0.809017im  -0.587785+0.809017im
  0.406737+0.913545im   0.406737+0.913545im   0.406737+0.913545im   0.406737+0.913545im   0.406737+0.913545im
  0.994522+0.104528im   0.994522+0.104528im   0.994522+0.104528im   0.994522+0.104528im   0.994522+0.104528im
  0.587785-0.809017im   0.587785-0.809017im   0.587785-0.809017im   0.587785-0.809017im   0.587785-0.809017im

SeparableFunctions.exp_ikx_lz — Function

exp_ikx_lz([::Type{TA},] sz::NTuple{N, Int}; pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

creates a Boolean box, being True inside and False outside. The side-length of the box can be defined via the argument boxsize.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

shift_by: a vector defining the real-space shift that would be caused by this function being multiplied in Fourier-space.

returns the a lazy version of an N-dimensional array only using memory for the separated 1-dimensional arrays. See the _col version for an example.

SeparableFunctions.exp_ikx_sep — Function

exp_ikx_sep([::Type{TA},] sz::NTuple{N, Int}; pos=zeros(eltype(TA),N), offset=sz.÷2 .+1, scale=1.0) where {TA, N}

yield an exp.(1im .* k .* Δx) function. A scaling of one means that the shift_by argument corresponds to Δx in integer pixels, but the term is in Fourier space. Note that the meaning of scale is different compared to the version in IndexFunArray.jl. Here the default scaling of 1.0 corresponds to the default scaling of ScaFT in IndexFunArrays.jl.To create the array on the GPU supply for example CuArray{Float32} as the array type. Note that the _col version yields a collected array, the _lz version a LazyArray and the _sep version an iterable of one-dimensional but oriented arrays, which can be used via the .*(res...) syntax with res being the result of the _sep call.

Arguments

TA: optionally an array type can be supplied. The default is Array{Float32}

sz: size of the result array

pos: position of the gaussian in relationship to the center defined by offset

offset: center position of the array

shift_by: a vector defining the real-space shift that would be caused by this function being multiplied in Fourier-space.

returns an iterable with the one-dimensional arrays, which can be used via .*(res...) wiht res being the result of calling this function.

–>

Radial functions

These functions are purely radial, in which case only a quadrant is computed and then copied. They are based on a separable view version of rr2 but then continue to calculate on a corner of a grid, which is (as a last step) replicated and mirrored.

Missing docstring.

Missing docstring for calc_radial_symm!. Check Documenter's build log for details.

SeparableFunctions.calc_radial_symm — Function

calc_radial_symm([::Type{TA},] sz::NTuple,  fct; scale = one(real(T)), myrr2sep = rr2_sep(size(arr); scale=scale, offset=size(arr).÷2 .+1)) where {N,T}

evaluates the radial function fct in a newly created array. The function needs to accept the square of the radius as argument! The calculation is done fast by only evaluating on the first quadrant and replicating the results by copy operations using copy_corners!().

#Arguments

TA: The array type for the newly created array.

sz: The size of the newly created array.

fct: The radial function of the squared radius, to be evaluate on the array coordinates.

scale: the vetorized scaling of the pixels (only used, if myrr2sep is not supplied by the user)

myrr2sep: The separable xx^2 and yy^2 etc. information as obtained by rr2_sep().

SeparableFunctions.propagator_col — Function

propagator_col([]::Type{TA},] sz::NTuple{N, Int}; Δz=one(eltype(TA)), k_max=0.5f0, scale=0.5f0 ./ (max.(sz ./ 2, 1))) where{TA, N}

generates a propagator for propagating optical fields via exp(i kz Δz) with kz=sqrt(k0^2-kx^2-ky^2). The k-space radius is stated by k_max relative to the Nyquist frequency, as long as the scale remains to be 1 ./ (2 max.(sz ./ 2, 1))).

#Arguments

TA: type of the array to generate. E.g. Array{Float64} or CuArray{Float32}.

sz: size of the array to generate. If a 3rd dimension is present, a stack a propagators is returned, one for each multiple of Δz.

Δz: distance in Z to propagate per slice.

k_max: maximum propagation radius in k-space. I.e. limit of the k-sphere. This is not the aperture limit!

scale: specifies how to interpret k-space positions. Should remain to be 1 ./ (2 max.(sz ./ 2, 1))).

SeparableFunctions.propagator_col! — Function

propagator_col(arr::AbstractArray{T,N}; Δz=one(eltype(TA)), k_max=0.5f0, scale=0.5f0 ./ (max.(sz ./ 2, 1))) where{TA, N}

generates a propagator for propagating optical fields via exp(i kz Δz) with kz=sqrt(k0^2-kx^2-ky^2). The k-space radius is stated by k_max relative to the Nyquist frequency, as long as the scale remains to be 1 ./ (2 max.(sz ./ 2, 1))).

#Arguments

arr: the array to fill with propagators. If a 3rd dimension is present, a stack a propagators is returned, one for each multiple of Δz.

Δz: distance in Z to propagate per slice.

k_max: maximum propagation radius in k-space. I.e. limit of the k-sphere. This is not the aperture limit!

scale: specifies how to interpret k-space positions. Should remain to be 1 ./ (2 max.(sz ./ 2, 1))).

SeparableFunctions.phase_kz_col — Function

phase_kz_col!(arr::AbstractArray{T,N}; Δz=one(eltype(TA)), k_max=0.5f0, scale=0.5f0 ./ (max.(sz ./ 2, 1))) where{TA, N}

Calculates a propagation phase (without the 2pi factor!) for a given z-position, which can be defined via a 3rd entry in the offset supplied to the function. By default, Nyquist sampling it is assumed such that the lateral kxy corresponds to the XY border in frequency space at the edge of the Ewald circle. However, via the xy scale entries the kmax can be set appropriately. The propagation equation uses offset[3] .* sqrt.(1-kxyrel^2) as the propagation phase. Therefore the Z-propagation distance (offset[3]) has to be specified in units of the wavelength in the medium (λ = n*λ₀). Note that since the phase is normalized to 1 instead of 2pi, you need to use this phase in the following sense: `cispi.(2.*phasekz(...))`.

#Arguments

arr: the array to fill with propagators. If a 3rd dimension is present, a stack a propagators is returned, one for each multiple of Δz.

Δz: distance in Z to propagate per slice.

k_max: maximum propagation radius in k-space. I.e. limit of the k-sphere. This is not the aperture limit!

scale: specifies how to interpret k-space positions. Should remain to be 1 ./ (2 max.(sz ./ 2, 1))).

Missing docstring.

Missing docstring for phase_kz_col!. Check Documenter's build log for details.